+++ front-end/mip/src/main/java/br/edu/utfpr/cp/emater/midmipsystem/service/mid/BladeReadingResponsibleEntityService.java
@@ -2,6 +2,7 @@ package br.edu.utfpr.cp.emater.midmipsystem.service.mid;
 
 import br.edu.utfpr.cp.emater.midmipsystem.entity.base.City;
 import br.edu.utfpr.cp.emater.midmipsystem.entity.mid.BladeReadingResponsibleEntity;
+import br.edu.utfpr.cp.emater.midmipsystem.entity.security.MIPUserPrincipal;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.AnyPersistenceException;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.EntityAlreadyExistsException;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.EntityInUseException;
@@ -11,23 +12,18 @@ import br.edu.utfpr.cp.emater.midmipsystem.service.ICRUDService;
 import br.edu.utfpr.cp.emater.midmipsystem.service.base.CityService;
 import java.util.ArrayList;
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.stereotype.Component;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
 
-@Component
+@Service
+@RequiredArgsConstructor
 public class BladeReadingResponsibleEntityService implements ICRUDService<BladeReadingResponsibleEntity> {
 
     private final BladeReadingResponsibleEntityRepository bladeEntityRepository;
     private final CityService cityService;
-
-    @Autowired
-    public BladeReadingResponsibleEntityService(BladeReadingResponsibleEntityRepository aBladeEntityRepository,
-            CityService aCityService) {
-
-        this.bladeEntityRepository = aBladeEntityRepository;
-        this.cityService = aCityService;
-    }
 
     @Override
     public List<BladeReadingResponsibleEntity> readAll() {
@@ -67,7 +63,7 @@ public class BladeReadingResponsibleEntityService implements ICRUDService<BladeR
         if (allPEntitiesWithoutExistentEntity.stream().anyMatch(current -> current.equals(aBladeEntity))) {
             throw new EntityAlreadyExistsException();
         }
-        
+
         var theCity = cityService.readById(aBladeEntity.getCityId());
 
         try {
@@ -84,6 +80,13 @@ public class BladeReadingResponsibleEntityService implements ICRUDService<BladeR
     public void delete(Long anId) throws EntityNotFoundException, EntityInUseException, AnyPersistenceException {
 
         var existentEntity = bladeEntityRepository.findById(anId).orElseThrow(EntityNotFoundException::new);
+
+        var loggedUser = ((MIPUserPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUser();
+        var createdByName = existentEntity.getCreatedBy() != null ? existentEntity.getCreatedBy().getUsername() : "none";
+
+        if (!loggedUser.getUsername().equalsIgnoreCase(createdByName)) {
+            throw new AccessDeniedException("Usuário não autorizado para essa exclusão!");
+        }
 
         try {
             bladeEntityRepository.delete(existentEntity);
