+++ front-end/mip/src/main/java/br/edu/utfpr/cp/emater/midmipsystem/service/mid/BladeReadingResponsiblePersonService.java
@@ -2,6 +2,7 @@ package br.edu.utfpr.cp.emater.midmipsystem.service.mid;
 
 import br.edu.utfpr.cp.emater.midmipsystem.entity.mid.BladeReadingResponsibleEntity;
 import br.edu.utfpr.cp.emater.midmipsystem.entity.mid.BladeReadingResponsiblePerson;
+import br.edu.utfpr.cp.emater.midmipsystem.entity.security.MIPUserPrincipal;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.AnyPersistenceException;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.EntityAlreadyExistsException;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.EntityInUseException;
@@ -10,23 +11,18 @@ import br.edu.utfpr.cp.emater.midmipsystem.repository.mid.BladeReadingResponsibl
 import br.edu.utfpr.cp.emater.midmipsystem.service.ICRUDService;
 import java.util.ArrayList;
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.stereotype.Component;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
 
-@Component
+@Service
+@RequiredArgsConstructor
 public class BladeReadingResponsiblePersonService implements ICRUDService<BladeReadingResponsiblePerson> {
 
     private final BladeReadingResponsiblePersonRepository bladePersonRepository;
     private final BladeReadingResponsibleEntityService bladeEntityService;
-
-    @Autowired
-    public BladeReadingResponsiblePersonService(BladeReadingResponsiblePersonRepository aBladePersonRepository,
-            BladeReadingResponsibleEntityService aBladeEntityService) {
-
-        this.bladePersonRepository = aBladePersonRepository;
-        this.bladeEntityService = aBladeEntityService;
-    }
 
     @Override
     public List<BladeReadingResponsiblePerson> readAll() {
@@ -66,7 +62,7 @@ public class BladeReadingResponsiblePersonService implements ICRUDService<BladeR
         if (allPersonsWithoutExistentPerson.stream().anyMatch(current -> current.equals(aBladePerson))) {
             throw new EntityAlreadyExistsException();
         }
-        
+
         var theEntity = bladeEntityService.readById(aBladePerson.getEntityId());
 
         try {
@@ -83,6 +79,13 @@ public class BladeReadingResponsiblePersonService implements ICRUDService<BladeR
     public void delete(Long anId) throws EntityNotFoundException, EntityInUseException, AnyPersistenceException {
 
         var existentPerson = bladePersonRepository.findById(anId).orElseThrow(EntityNotFoundException::new);
+
+        var loggedUser = ((MIPUserPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUser();
+        var createdByName = existentPerson.getCreatedBy() != null ? existentPerson.getCreatedBy().getUsername() : "none";
+
+        if (!loggedUser.getUsername().equalsIgnoreCase(createdByName)) {
+            throw new AccessDeniedException("Usuário não autorizado para essa exclusão!");
+        }
 
         try {
             bladePersonRepository.delete(existentPerson);
