+++ front-end/mip/src/main/java/br/edu/utfpr/cp/emater/midmipsystem/service/mid/MIDRustSampleService.java
@@ -2,6 +2,7 @@ package br.edu.utfpr.cp.emater.midmipsystem.service.mid;
 
 import br.edu.utfpr.cp.emater.midmipsystem.entity.mid.BladeReadingResponsiblePerson;
 import br.edu.utfpr.cp.emater.midmipsystem.entity.mid.MIDRustSample;
+import br.edu.utfpr.cp.emater.midmipsystem.entity.security.MIPUserPrincipal;
 import br.edu.utfpr.cp.emater.midmipsystem.entity.survey.Survey;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.AnyPersistenceException;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.EntityAlreadyExistsException;
@@ -11,28 +12,21 @@ import br.edu.utfpr.cp.emater.midmipsystem.repository.mid.MIDRustSampleRepositor
 import br.edu.utfpr.cp.emater.midmipsystem.service.survey.SurveyService;
 import java.util.List;
 import java.util.stream.Collectors;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.stereotype.Component;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
 
-@Component
+@Service
+@RequiredArgsConstructor
 public class MIDRustSampleService {
 
     private final MIDRustSampleRepository midRustSampleRepository;
 
     private final SurveyService surveyService;
-    
+
     private final BladeReadingResponsiblePersonService bladeResponsiblePersonService;
-
-    @Autowired
-    public MIDRustSampleService(MIDRustSampleRepository midRustSampleRepository,
-            SurveyService aSurveyService,
-            BladeReadingResponsiblePersonService aBladeResponsiblePersonService) {
-
-        this.midRustSampleRepository = midRustSampleRepository;
-        this.surveyService = aSurveyService;
-        this.bladeResponsiblePersonService = aBladeResponsiblePersonService;
-    }
 
     public Survey readSurveyById(Long id) throws EntityNotFoundException {
         return surveyService.readById(id);
@@ -65,9 +59,16 @@ public class MIDRustSampleService {
         return List.copyOf(this.midRustSampleRepository.findAll().stream().filter(current -> current.getSurvey().getId().equals(id)).collect(Collectors.toList()));
     }
 
-    public void delete(Long anId) throws EntityNotFoundException, EntityInUseException, AnyPersistenceException{
+    public void delete(Long anId) throws EntityNotFoundException, EntityInUseException, AnyPersistenceException {
 
         var existentSample = midRustSampleRepository.findById(anId).orElseThrow(EntityNotFoundException::new);
+
+        var loggedUser = ((MIPUserPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUser();
+        var createdByName = existentSample.getCreatedBy() != null ? existentSample.getCreatedBy().getUsername() : "none";
+
+        if (!loggedUser.getUsername().equalsIgnoreCase(createdByName)) {
+            throw new AccessDeniedException("Usuário não autorizado para essa exclusão!");
+        }
 
         try {
             midRustSampleRepository.delete(existentSample);
