+++ front-end/mip/src/main/java/br/edu/utfpr/cp/emater/midmipsystem/service/pulverisation/PulverisationOperationService.java
@@ -1,45 +1,35 @@
 package br.edu.utfpr.cp.emater.midmipsystem.service.pulverisation;
 
-import br.edu.utfpr.cp.emater.midmipsystem.service.survey.*;
 import br.edu.utfpr.cp.emater.midmipsystem.entity.pulverisation.Product;
 import br.edu.utfpr.cp.emater.midmipsystem.entity.pulverisation.PulverisationOperation;
 import br.edu.utfpr.cp.emater.midmipsystem.entity.pulverisation.Target;
-import br.edu.utfpr.cp.emater.midmipsystem.entity.pulverisation.TargetCategory;
+import br.edu.utfpr.cp.emater.midmipsystem.entity.pulverisation.UseClass;
+import br.edu.utfpr.cp.emater.midmipsystem.entity.security.MIPUserPrincipal;
 import br.edu.utfpr.cp.emater.midmipsystem.entity.survey.Survey;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.AnyPersistenceException;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.EntityAlreadyExistsException;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.EntityInUseException;
 import br.edu.utfpr.cp.emater.midmipsystem.exception.EntityNotFoundException;
 import br.edu.utfpr.cp.emater.midmipsystem.repository.pulverisation.PulverisationOperationRepository;
+import br.edu.utfpr.cp.emater.midmipsystem.service.survey.SurveyService;
 import java.util.Date;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.stereotype.Component;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
 
-@Component
+@Service
+@RequiredArgsConstructor
 public class PulverisationOperationService {
-    
+
     private final PulverisationOperationRepository pulverisationOperationRepository;
     private final TargetService targetService;
     private final ProductService productService;
     private final SurveyService surveyService;
-
-    @Autowired
-    public PulverisationOperationService(
-            PulverisationOperationRepository aPulverisationOperationRepository,
-            TargetService aTargetService,
-            ProductService aProductService,
-            SurveyService aSurveyService) {
-
-        this.pulverisationOperationRepository = aPulverisationOperationRepository;
-        this.targetService = aTargetService;
-        this.productService = aProductService;
-        this.surveyService = aSurveyService;
-    }
-
 
     public List<Survey> readAllSurveysUniqueEntries() {
         return List.copyOf(pulverisationOperationRepository.findAll().stream().map(PulverisationOperation::getSurvey).distinct().collect(Collectors.toList()));
@@ -50,8 +40,6 @@ public class PulverisationOperationService {
         if (pulverisationOperationRepository.findAll().stream().anyMatch(current -> current.equals(anOperation))) {
             throw new EntityAlreadyExistsException();
         }
-        
-        anOperation.setDaysAfterEmergence(this.calculateDaysAfterEmergence(anOperation.getSurvey().getEmergenceDate(), anOperation.getSampleDate()));
 
         try {
             pulverisationOperationRepository.save(anOperation);
@@ -62,23 +50,21 @@ public class PulverisationOperationService {
 
         }
     }
-    
-    private int calculateDaysAfterEmergence(Date emergenceDate, Date sampleDate) {
-
-        long diffInMillies = Math.abs(sampleDate.getTime() - emergenceDate.getTime());
-        
-        var result = TimeUnit.DAYS.convert(diffInMillies, TimeUnit.MILLISECONDS);
-        
-        return (int) (result + 1);
-    }
 
     public PulverisationOperation readById(Long anId) throws EntityNotFoundException {
         return pulverisationOperationRepository.findById(anId).orElseThrow(EntityNotFoundException::new);
     }
 
     public void delete(Long anId) throws EntityNotFoundException, EntityInUseException, AnyPersistenceException {
-        
+
         var existentOperation = pulverisationOperationRepository.findById(anId).orElseThrow(EntityNotFoundException::new);
+
+        var loggedUser = ((MIPUserPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUser();
+        var createdByName = existentOperation.getCreatedBy() != null ? existentOperation.getCreatedBy().getUsername() : "none";
+
+        if (!loggedUser.getUsername().equalsIgnoreCase(createdByName)) {
+            throw new AccessDeniedException("Usuário não autorizado para essa exclusão!");
+        }
 
         try {
             pulverisationOperationRepository.delete(existentOperation);
@@ -100,12 +86,12 @@ public class PulverisationOperationService {
         return List.copyOf(pulverisationOperationRepository.findAll().stream().filter(current -> current.getSurvey().getId().equals(aSurveyId)).collect(Collectors.toList()));
     }
 
-    public List<Target> readAllTargetsByCategory(TargetCategory targetCategory) {
-        return targetService.readAll().stream().filter(current -> current.getCategory().equals(targetCategory)).collect(Collectors.toList());
+    public List<Target> readAllTargetsByUseClass(UseClass aUseClass) {
+        return targetService.readAll().stream().filter(current -> current.getUseClass().equals(aUseClass)).collect(Collectors.toList());
     }
 
-    public List<Product> readAllProductByTarget(Long targetId) {
-        return productService.readAll().stream().filter(current -> current.getTargetId().equals(targetId)).collect(Collectors.toList());
+    public List<Product> readAllProductByUseClass(UseClass aUseClass) {
+        return productService.readAll().stream().filter(current -> current.getUseClass().equals(aUseClass)).collect(Collectors.toList());
     }
 
     public Product readProductById(Long productId) throws EntityNotFoundException {
